[Refactoring] Move console related classes to package
