+++ src/main/kotlin/com/elpassion/intelijidea/common/console/MFCommandLineState.kt
@@ -1,4 +1,4 @@
-package com.elpassion.intelijidea.common
+package com.elpassion.intelijidea.common.console
 
 import com.intellij.execution.ExecutionManager
 import com.intellij.execution.ExecutionResult
